+++ ecosystem/plugin-docsearch/src/client/composables/index.ts
@@ -1,1 +1,2 @@
+export * from './useDocsearchHotkeyListener.js'
 export * from './useDocsearchShim.js'