+++ packages/client/src/components/Vuepress.ts
@@ -1,19 +1,5 @@
-import { clientConfigs } from '@internal/clientConfigs'
-import { isString } from '@vuepress/shared'
-import type { Component } from 'vue'
-import { computed, defineComponent, h } from 'vue'
-import { usePageData } from '../composables/index.js'
-
-const LAYOUT_NAME_DEFAULT = 'Layout'
-const LAYOUT_NAME_NOT_FOUND = 'NotFound'
-
-const layouts = clientConfigs.reduce(
-  (prev, item) => ({
-    ...prev,
-    ...item.layouts,
-  }),
-  {} as Record<string, Component>
-)
+import { defineComponent, h } from 'vue'
+import { usePageLayout } from '../composables/index.js'
 
 /**
  * Global Layout
@@ -22,33 +8,7 @@ export const Vuepress = defineComponent({
   name: 'Vuepress',
 
   setup() {
-    const page = usePageData()
-
-    // resolve layout component
-    const layoutComponent = computed(() => {
-      // resolve layout name of current page
-      let layoutName: string
-
-      if (page.value.path) {
-        // if current page exists
-
-        // use layout from frontmatter
-        const frontmatterLayout = page.value.frontmatter.layout
-
-        if (isString(frontmatterLayout)) {
-          layoutName = frontmatterLayout
-        } else {
-          // fallback to default layout
-          layoutName = LAYOUT_NAME_DEFAULT
-        }
-      } else {
-        // if current page does not exist
-        // use NotFound layout
-        layoutName = LAYOUT_NAME_NOT_FOUND
-      }
-      return layouts[layoutName]
-    })
-
-    return () => h(layoutComponent.value)
+    const layout = usePageLayout()
+    return () => h(layout.value)
   },
 })