+++ packages/client/src/setupUpdateHead.ts
@@ -1,7 +1,6 @@
 import { isPlainObject, isString } from '@vuepress/shared'
 import type { HeadConfig, VuepressSSRContext } from '@vuepress/shared'
 import { onMounted, provide, ref, useSSRContext, watch } from 'vue'
-import { useRoute } from 'vue-router'
 import {
   updateHeadSymbol,
   usePageHead,
@@ -13,7 +12,6 @@ import type { UpdateHead } from './composables/index.js'
  * Auto update head and provide as global util
  */
 export const setupUpdateHead = (): void => {
-  const route = useRoute()
   const head = usePageHead()
   const lang = usePageLang()
 
@@ -64,11 +62,7 @@ export const setupUpdateHead = (): void => {
     loadHead()
     updateHead()
     watch(
-      // when watching `head`, route hash changes will also trigger the watcher,
-      // causing unnecessary head updates
-      // so we watch `head` in dev mode to support hot-reload of `frontmatter.head`,
-      // and watch `route.path` in production mode to avoid extra updates
-      () => (__VUEPRESS_DEV__ ? head.value : route.path),
+      () => head.value,
       () => updateHead()
     )
   })
@@ -85,7 +79,7 @@ export const queryHeadTag = ([
   const attrsSelector = Object.entries(attrs)
     .map(([key, value]) => {
       if (isString(value)) {
-        return `[${key}="${value}"]`
+        return `[${key}=${JSON.stringify(value)}]`
       }
       if (value === true) {
         return `[${key}]`