+++ packages/client/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './clientConfig.js'
 export * from './createVueAppFunction.js'
+export * from './layouts.js'
 export * from './pageRouteItem.js'