+++ packages/client/src/types/layouts.ts
@@ -1,0 +1,3 @@
+import type { Component } from 'vue'
+
+export type Layouts = Record<string, Component>